aification_backend/requirements.txt
# Enhanced requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic[email]>=2.5.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Google authentication
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Additional security and utilities
python-multipart>=0.0.6  # For form data
email-validator>=2.1.0   # For email validation
httpx>=0.25.0           # For HTTP requests (optional)

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.6.0